Fix: Voucher note online, input giá chung, tìm kiếm hóa đơn online

+++ be/src/main/java/com/sd20201/datn/core/admin/products/product/model/request/ADProductRequest.java
@@ -16,8 +16,8 @@ public class ADProductRequest extends PageableRequest {
 
     private String idOperatingSystem;
 
-    private Integer minPrice;
+    private Long minPrice;
 
-    private Integer maxPrice;
+    private Long maxPrice;
 
 }
