+++ be/src/main/java/com/sd20201/datn/core/admin/products/product/model/response/ADProductResponse.java
@@ -17,9 +17,9 @@ public interface ADProductResponse extends IsIdentify, IsEntityStatus {
 
     String getScreen();
 
-    Integer getMinPrice();
+    Long getMinPrice();
 
-    Integer getMaxPrice();
+    Long getMaxPrice();
 
     Integer getQuantity();
